+++ TP2/main.c
@@ -4,170 +4,147 @@
 #include <stdint.h>
 #include <math.h>
 
-typedef struct s_mesh { // 5.2 Projection and view matrices
-    uint32_t nVertices, nTriangles;
-    float vertices[8][3]; // 8 sommets définis sur 3 coordonnées
-    uint32_t triangles[12][3]; // 12 triangles de 3 sommets
-} Cube;
+#define NB_VERTICES_CUBE 8
+#define NB_TRIANGLES_CUBE 12
 
-Cube MY_CUBE;
+/* Structures */
+typedef struct mesh{
+    uint8_t    nb_vertices;
+    float       *vertices;
+    
+    uint8_t    nb_triangles;
+    uint8_t    *triangles; 
 
-
-typedef struct { // 5.2 Projection and view matrices
-    double ox, oy, oz; // Position de la caméra
-    double fx, fy, fz; // Focus
-    double vx, vy, vz; // Up vector
-} ViewMatrix;
-typedef struct { // 5.2 Projection and view matrices
-    double left, right;
-    double bottom, top;
-    double nearVal, farVal;
-} ProjectionMatrix;
+} Mesh ;
 
 
-ViewMatrix getDefaultCamera() { // 5.2 Projection and view matrices
-    ViewMatrix cam = { // utilisation de la syntaxe c99
-        .ox = 2, .oy = 2, .oz = 2,    // Position de la caméra
-        .fx = 0, .fy = 0, .fz = 0,    // Cible 
-        .vx = 0, .vy = 1, .vz = 0     // "Up" vector
-    };
-    return cam;
-}
-ProjectionMatrix getDefaultProjection() { // 5.2 Projection and view matrices
-    ProjectionMatrix proj = {
-        .left = -1.0, .right = 1.0,
-        .bottom = -1.0, .top = 1.0,
-        .nearVal = 1.0, .farVal = 10.0
-    };
-    return proj;
-}
+/* Variables globales */
+Mesh msh;
 
 
 
-void viewMatrix(){ // 5.2 Projection and view matrices
-    ViewMatrix camera = getDefaultCamera();
-    glMatrixMode(GL_MODELVIEW);
-    glLoadIdentity();
-    gluLookAt(
-        camera.ox, camera.oy, camera.oz,
-        camera.fx, camera.fy, camera.fz,
-        camera.vx, camera.vy, camera.vz
-    );
-}
-void projectionMatrix(){ // 5.2 Projection and view matrices
-    ProjectionMatrix projection = getDefaultProjection();
-    glMatrixMode(GL_PROJECTION);
-    glLoadIdentity();
-    glFrustum(
-        projection.left, projection.right,
-        projection.bottom, projection.top,
-        projection.nearVal, projection.farVal
-    );
-}
-
-
-
-void initCube(){
-
-    MY_CUBE.nVertices = 8;
-    MY_CUBE.nTriangles = 12;
-/* Merci ChatGPT
- A = {0,0,0}
-    E--------F
-    /|       /|
-   H--------G |
-   | |      | |
-   | A------|-B
-   |/       |/
-   D--------C
+void defineCube(void){
+    uint8_t i;
+    
+    
+    msh.nb_vertices = NB_VERTICES_CUBE;
+    msh.vertices    = (float*)malloc(msh.nb_vertices*3*sizeof(float));
+/*    
+          4-----------5
+         /|          /|
+        7-----------6 |
+        | |         | |
+        | 0---------|-1
+        |/          |/
+        3-----------2---- X
 */
-
-    // On peut faire ca comme ca pour cette fois car le cube n'a que 8 coins
-    // En général, on évitera de stocker des grands tableaux dans la stack
-    // 8 points * 3 dimensions
-    float points[] = {
-        0, 0, 0, // A (0)
-        0, 1, 0, // B (1)
-        1, 1, 0, // C (2)
-        1, 0, 0, // D (3)
-        0, 0, 1, // E (4)
-        0, 1, 1, // F (5)
-        1, 1, 1, // G (6)
-        1, 0, 1  // H (7)
-    };
-    for (uint8_t i = 0; i < 8; i++) {
-        MY_CUBE.vertices[i][0] = points[i * 3 + 0];
-        MY_CUBE.vertices[i][1] = points[i * 3 + 1];
-        MY_CUBE.vertices[i][2] = points[i * 3 + 2];
+    for (i=0; i<msh.nb_vertices*3;i++ ){
+        msh.vertices[i]=0.0;
     }
+    
+    // Bottom
+    // 0
+    i=0;
+    // 1
+    i++;
+    msh.vertices[3*i +1 ]=1.0;
+    // 2
+    i++;
+    msh.vertices[3*i +1 ]=1.0;
+    msh.vertices[3*i ]=1.0;
+    // 3
+    i++;
+    msh.vertices[3*i  ]=1.0;
+    // Top
+    // 4
+    i++;
+    msh.vertices[3*i +2 ]=1.0;
+    // 5
+    i++;
+    msh.vertices[3*i +1 ]=1.0;
+    msh.vertices[3*i +2 ]=1.0;
+    // 6
+    i++;
+    msh.vertices[3*i  ]=1.0;
+    msh.vertices[3*i +1 ]=1.0;
+    msh.vertices[3*i +2 ]=1.0;
+    // 7
+    i++;
+    msh.vertices[3*i ]=1.0;
+    msh.vertices[3*i +2 ]=1.0;
 
-    // 12 triangles * 3 indices
-    uint32_t triangles[] = {
-        0, 1, 2,  // face avant
-        2, 3, 0,
-        3, 7, 2,  // face droite
-        7, 6, 2,
-        1, 2, 6,  // face haut
-        6, 5, 1,
-        4, 5, 6,  // face arrière
-        6, 7, 4,
-        0, 3, 7,  // face bas
-        7, 4, 0,
-        0, 1, 5,  // face gauche
-        5, 4, 0
-    };
-    for (int i = 0; i < 12; i++) {
-        MY_CUBE.triangles[i][0] = triangles[i * 3 + 0];
-        MY_CUBE.triangles[i][1] = triangles[i * 3 + 1];
-        MY_CUBE.triangles[i][2] = triangles[i * 3 + 2];
-    }
+    msh.nb_triangles = NB_TRIANGLES_CUBE;
+    msh.triangles=(uint8_t*)malloc(msh.nb_triangles*3*sizeof(uint8_t));
+
+    
+    // Face du bas
+    i=0;
+    msh.triangles[i*3 + 0]=0;
+    msh.triangles[i*3 + 1]=1;
+    msh.triangles[i*3 + 2]=2;
+    i++;
+    msh.triangles[i*3 + 0]=0;
+    msh.triangles[i*3 + 1]=2;
+    msh.triangles[i*3 + 2]=3;
+    // Face de gauche
+    i++;
+    msh.triangles[i*3 + 0]=7;
+    msh.triangles[i*3 + 1]=4;
+    msh.triangles[i*3 + 2]=0;
+    i++;
+    msh.triangles[i*3 + 0]=0;
+    msh.triangles[i*3 + 1]=3;
+    msh.triangles[i*3 + 2]=7;
+    // Face de droite
+    i++;
+    msh.triangles[i*3 + 0]=1;
+    msh.triangles[i*3 + 1]=5;
+    msh.triangles[i*3 + 2]=6;
+    i++;
+    msh.triangles[i*3 + 0]=6;
+    msh.triangles[i*3 + 1]=2;
+    msh.triangles[i*3 + 2]=1;
+    // Face du haut
+    i++;
+    msh.triangles[i*3 + 0]=6;
+    msh.triangles[i*3 + 1]=5;
+    msh.triangles[i*3 + 2]=4;
+    i++;
+    msh.triangles[i*3 + 0]=4;
+    msh.triangles[i*3 + 1]=7;
+    msh.triangles[i*3 + 2]=6;
+    // Face de devant
+    i++;
+    msh.triangles[i*3 + 0]=3;
+    msh.triangles[i*3 + 1]=2;
+    msh.triangles[i*3 + 2]=7;
+    i++;
+    msh.triangles[i*3 + 0]=7;
+    msh.triangles[i*3 + 1]=2;
+    msh.triangles[i*3 + 2]=6;
+    // Face de derriere
+    i++;
+    msh.triangles[i*3 + 0]=1;
+    msh.triangles[i*3 + 1]=0;
+    msh.triangles[i*3 + 2]=5;
+    i++;
+    msh.triangles[i*3 + 0]=0;
+    msh.triangles[i*3 + 1]=4;
+    msh.triangles[i*3 + 2]=5;
 }
 
 
-void displayCube(){
-    glClear(GL_COLOR_BUFFER_BIT);  // Clear the screen
-
-    glColor3f(1, 0.5, 0.5);
-    glPolygonMode(GL_FRONT_AND_BACK, GL_LINE); // Draw only lines (wireframe)
-    // For each Triangle
-    for (uint8_t triangle = 0; triangle < MY_CUBE.nTriangles; triangle++) {
-        glBegin(GL_TRIANGLES);
-        // For each point of the Triangle
-        for (uint8_t point_index = 0; point_index < 3; point_index++) {
-            uint32_t vertex_index = MY_CUBE.triangles[triangle][point_index];
-            float x = MY_CUBE.vertices[vertex_index][0];
-            float y = MY_CUBE.vertices[vertex_index][1];
-            float z = MY_CUBE.vertices[vertex_index][2];
-            printf("Plotting point %d of triangle %d at (%f, %f, %f)...\n",
-                   vertex_index, triangle, x, y, z);
-            glVertex3f(x, y, z);
-        }
-        glEnd();
-    }
-    glFlush();  
-}
 
 
-void initColors() {
-    glClearColor(1.0, 1.0, 1.0, 1.0);
-    glColor3f(0.0, 0.0, 0.0);
-}
 
-void initOpengGL(int argc, char **argv) {
-    glutInit(&argc, argv);
-    glutInitDisplayMode(GLUT_SINGLE | GLUT_RGB);
-    glutInitWindowSize(1000, 1000);
-    glutInitWindowPosition(10, 10);
-    glutCreateWindow("TPGL");
-    glutDisplayFunc(displayCube);
-}
+int main(int argc, uint8_t **argv){
+    InitialiserLibrairieGraphique(&argc,argv); // -> glutInit
+    InitialiserParametresGraphiques();
+    ModeleDiscret();
+    CreationFenetreGraphique();
+    EvenementsGraphiques();
+    BoucleInfinie();
+    
 
-int main(int argc, char **argv) {
-    initOpengGL(argc, argv);
-    initCube();
-    initColors();
-    viewMatrix();
-    projectionMatrix();
-    glutMainLoop();
-    return EXIT_SUCCESS;
+    return 1;
 }
